refactor: move token rotation utilities
